+++ helper.py
@@ -120,7 +120,7 @@ def GaussianPDF_2D(mu, sigma, row, col):
 def rgb2gray(I_rgb):
   r, g, b = I_rgb[:, :, 0], I_rgb[:, :, 1], I_rgb[:, :, 2]
   I_gray = 0.2989 * r + 0.5870 * g + 0.1140 * b
-  return I_gray
+  return I_gray.astype(np.uint8)
